+++ spray/src/main/scala/com/monsanto/labs/mwundo/GeoJsonFormats.scala
@@ -18,7 +18,7 @@ object GeoJsonFormats extends DefaultJsonProtocol {
 
     def read(json: JsValue): Coordinate = json match {
       case JsArray(is) if is.length == 2 =>
-        Coordinate(is(0).convertTo[BigDecimal], is(1).convertTo[BigDecimal])
+        Coordinate(is(0).convertTo[Double], is(1).convertTo[Double])
       case _ => deserializationError(s"'$json' is not a valid Coordinate")
     }
   }
